refactor(linkedlist): Node를 LinkedList의 private inner class로 변경
